Add files via upload
We work in a team of two, me and Dean Meyer, to create a Downloader platform that can speed-up the download of a file using multiple HTTP connections, multiple servers, and recovering from a previously stopped download using metadata. It can speed up to a few minutes the download of a big file.